Fix compilation errors: update ServiceLineItemCommand to use Double
Updated ServiceLineItemCommand.adjustmentValue from Long to Double to match
the updated request DTOs (PriceAdjustmentRequest).

Also added conversion logic in GetAppointmentHandler to convert basis points
back to percentages when returning appointment details (same as was done for
ListAppointmentsHandler).

Fixes:
- AppointmentController.kt: type mismatch resolved
- GetAppointmentHandler.kt: type mismatch resolved and conversion added
- ServiceLineItemCommand: adjustmentValue changed from Long to Double

+++ src/main/kotlin/pl/detailing/crm/appointment/create/CreateAppointmentCommand.kt
@@ -66,7 +66,7 @@ data class ServiceLineItemCommand(
     val serviceId: ServiceId?,
     val serviceName: String?,
     val adjustmentType: AdjustmentType,
-    val adjustmentValue: Long,
+    val adjustmentValue: Double,  // Double to support decimal percentages
     val customNote: String?
 )
